Skip assertions in `run_test!`
Create a new method called `run_test_fake!` that performs the OpenAPI
spec generation of a particular response, but without checking the
test assertions. This is useful when we want to generate a custom
response in the spec that does not follow the actual API
behaviour (e.g. when dealing with redirects to resources served by
other services).